+++ Sources/RKAPIService/HTTPHeader.swift
@@ -8,15 +8,43 @@
 import Foundation
 
 ///HTTPHeader for URLRequest
-public struct HTTPHeader {
+public struct HTTPHeader: Header {
     ///HTTPHeader key
-    let key: String
+    public let key: String
     
     ///HTTPHeader value
-    let value: String
+    public let value: String
     
     public init(key: String, value: String) {
         self.key = key
         self.value = value
+    }
+}
+
+///Header for URLRequest
+ public protocol Header {
+    ///Header key
+    var key: String {get}
+    
+    ///Header value
+    var value: String {get}
+}
+
+public enum Authorization: Header {
+    case bearerToken(token: String)
+    
+    public var key: String {
+        get {
+            return "Authorization"
+        }
+    }
+    
+    public var value: String {
+        get {
+            switch self {
+                case .bearerToken(let token):
+                    return "Bearer \(token)"
+            }
+        }
     }
 }