+++ Sources/RKAPIService/RKAPIService.swift
@@ -4,12 +4,15 @@ import Combine
 /**
  RKAPIService class. It implements `RKAPIServiceProtocol`. This class serves all the necessary steps to perform a `URLSession` call.
  */
-public class RKAPIService: RKAPIServiceProtocol {
+public class RKAPIService {
     
     /// Static instance of `RKAPIService`. It has `URLSessionConfiguration.ephemeral` as configuration. `URLSessionDelegate` and `OperationQueue` are both nil.
-    public static var shared = RKAPIService(sessionConfiguration: URLSessionConfiguration.ephemeral, delegate: nil, queue: nil)
+    public static var shared = RKAPIService()
     
-    private let session: URLSession
+    private var session: URLSession
+    private var config: URLSessionConfiguration
+    private var delegate: URLSessionDelegate?
+    private var queue: OperationQueue?
     
     /**
      Initializes ``RKAPIService``
@@ -19,9 +22,11 @@ public class RKAPIService: RKAPIServiceProtocol {
         - delegate: Receives an `Optional<URLSessionDelegate>` or `URLSessionDelegate?` from `Foundation`
         - queue: Receiives an `Optional<OperationQueue>` or `OperationQueue?` from `Foundation`
      */
-    public init(sessionConfiguration: URLSessionConfiguration, delegate: URLSessionDelegate?, queue: OperationQueue?) {
-        
-        session = URLSession(configuration: sessionConfiguration, delegate: delegate, delegateQueue: queue)
+    public init(sessionConfiguration: URLSessionConfiguration = .ephemeral, delegate: URLSessionDelegate? = nil, queue: OperationQueue? = nil) {
+        self.config = sessionConfiguration
+        self.delegate = delegate
+        self.queue = queue
+        self.session = URLSession(configuration: sessionConfiguration, delegate: delegate, delegateQueue: queue)
     }
     
     /**
@@ -31,12 +36,46 @@ public class RKAPIService: RKAPIServiceProtocol {
         session.invalidateAndCancel()
     }
     
+    /**
+     Replaces the current session with a new one
+     
+     If at any point we need  to update our session then we call this method. If we pass the parameters then it will update session with new values. By default it will just reset the session.
+     
+     - Parameters:
+        - sessionConfiguration: Receives `Optional<URLSessionConfiguration>` or `URLSessionConfiguration?` from `Foundation`
+        - delegate: Receives an `Optional<URLSessionDelegate>` or `URLSessionDelegate?` from `Foundation`
+        - queue: Receiives an `Optional<OperationQueue>` or `OperationQueue?` from `Foundation`
+     */
+    public func invalidateAndReinitializeSession(sessionConfiguration: URLSessionConfiguration? = nil, delegate: URLSessionDelegate? = nil, queue: OperationQueue? = nil) {
+        invalidateAndCancelSession()
+        
+        var actualConfig: URLSessionConfiguration = self.config
+        var actualDelegate: URLSessionDelegate? = self.delegate
+        var actualQueue: OperationQueue? = self.queue
+        
+        if let sessionConfiguration = sessionConfiguration {
+            actualConfig = sessionConfiguration
+        }
+        
+        if let delegate = delegate {
+            actualDelegate = delegate
+        }
+        
+        if let queue = queue {
+            actualQueue = queue
+        }
+        
+        let newSession = URLSession(configuration: actualConfig, delegate: actualDelegate, delegateQueue: actualQueue)
+        
+        session = newSession
+    }
+    
     @available(iOS 13.0, macOS 10.15.0, *)
     internal func previousVersionURLSession(request: URLRequest) async throws -> (Data, URLResponse) {
         try await withCheckedThrowingContinuation({ (continuation: CheckedContinuation<(Data, URLResponse), Error>) in
             session.dataTask(with: request) { data, response, error in
                 if let error = error {
-                 continuation.resume(throwing: error)
+                    continuation.resume(throwing: error)
                 }
                 else {
                     if let data = data, let response = response {
@@ -48,7 +87,7 @@ public class RKAPIService: RKAPIServiceProtocol {
     }
 }
 
-//MARK: With async/await
+//MARK: - With async/await
 @available(iOS 13.0, macOS 10.15.0, *)
 extension RKAPIService {
     /**
@@ -56,13 +95,13 @@ extension RKAPIService {
      
      - Parameters:
         - urlLink: Receives an `Optional<URL>` aka `URL?`
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
      
      - Throws: An `URLError` is thrown if urlLink is nil or not a valied URL or server does not provide any response. Also ``HTTPStatusCode`` Error (Custom error) can be thrown if server status code is anything but 200...299
      
      - Returns: Returns a  ``NetworkResult``
      */
-    public func fetchItems(urlLink: URL?, additionalHeader: [HTTPHeader]? = nil) async throws -> NetworkResult<Data> {
+    public func fetchItems(urlLink: URL?, additionalHeader: [Header]? = nil) async throws -> NetworkResult<Data> {
         guard let url = urlLink else {
             throw URLError(.badURL)
         }
@@ -102,12 +141,12 @@ extension RKAPIService {
      
      - Parameters:
         - urlLink: Receives an `Optional<URL>` aka `URL?`
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - model: Generic Type `D` where `D` confirms to `Decodable`
      
      - Returns: Returns a  `Result<Success, Failure>` type where `Success` is  ``NetworkResult`` and failure is `Error`
      */
-    public func fetchItems<D: Decodable>(urlLink: URL?, additionalHeader: [HTTPHeader]? = nil, _ model: D.Type) async -> Result<NetworkResult<D>, Error> {
+    public func fetchItems<D: Decodable>(urlLink: URL?, additionalHeader: [Header]? = nil, _ model: D.Type) async -> Result<NetworkResult<D>, Error> {
         do {
             let reply = try await fetchItems(urlLink: urlLink, additionalHeader: additionalHeader)
             
@@ -121,6 +160,7 @@ extension RKAPIService {
         }
     }
     
+    //MARK: - Original fetchItemsByHTTPMethod
     /**
      Fetch items with HTTP method.
      
@@ -130,13 +170,13 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: `Optional<Data>` aka `Data?` for sending to remote server.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
      
      - Throws: An `URLError` is thrown if urlLink is nil or not a valied `URL` or server does not provide any response. Also ``HTTPStatusCode`` Error (Custom error) can be thrown if server status code is anything but 200...299
      
      - Returns: Returns a  ``NetworkResult``
      */
-    public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, additionalHeader: [HTTPHeader]? = nil) async throws -> NetworkResult<Data> {
+    public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, additionalHeader: [Header]? = nil) async throws -> NetworkResult<Data> {
         guard let url = urlLink else {
             throw URLError(.badURL)
         }
@@ -184,12 +224,12 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: `Optional<Data>` aka `Data?` for sending to remote server.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - model: Generic Type `D` where `D` confirms to `Decodable`
      
      - Returns: Returns a  `Result<Success, Failure>` type where `Success` is  ``NetworkResult`` and failure is `Error`
      */
-    public func fetchItemsByHTTPMethod<D: Decodable>(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, additionalHeader: [HTTPHeader]? = nil, _ model: D.Type) async -> Result<NetworkResult<D>, Error> {
+    public func fetchItemsByHTTPMethod<D: Decodable>(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, additionalHeader: [Header]? = nil, _ model: D.Type) async -> Result<NetworkResult<D>, Error> {
         do {
             let reply = try await fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: body, additionalHeader: additionalHeader)
             
@@ -212,11 +252,11 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: Generic Type `E` where `E` confirms to `Encodable`.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
      
      - Returns: Returns a  `Result<Success, Failure>` type where `Success` is  ``NetworkResult`` and failure is `Error`
      */
-    public func fetchItemsByHTTPMethod<E: Encodable>(urlLink: URL?, httpMethod: HTTPMethod, body: E, additionalHeader: [HTTPHeader]? = nil) async throws -> NetworkResult<Data> {
+    public func fetchItemsByHTTPMethod<E: Encodable>(urlLink: URL?, httpMethod: HTTPMethod, body: E, additionalHeader: [Header]? = nil) async throws -> NetworkResult<Data> {
         let uploadData = RKAPIHelper.generateRequestBody(body)
         let reply = try await fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: uploadData, additionalHeader: additionalHeader)
         
@@ -232,20 +272,72 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: Generic Type `E` where `E` confirms to `Encodable`.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - model: Generic Type `D` where `D` confirms to `Decodable`.
      
      - Returns: Returns a  `Result<Success, Failure>` type where `Success` is  ``NetworkResult`` and failure is `Error`
      */
-    public func fetchItemsByHTTPMethod<D: Decodable, E: Encodable>(urlLink: URL?, httpMethod: HTTPMethod, body: E, additionalHeader: [HTTPHeader]? = nil, _ model: D.Type) async -> Result<NetworkResult<D>, Error> {
+    public func fetchItemsByHTTPMethod<D: Decodable, E: Encodable>(urlLink: URL?, httpMethod: HTTPMethod, body: E, additionalHeader: [Header]? = nil, _ model: D.Type) async -> Result<NetworkResult<D>, Error> {
         let uploadData = RKAPIHelper.generateRequestBody(body)
         let reply = await fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: uploadData, additionalHeader: additionalHeader, D.self)
         
         return reply
     }
+    
+    //MARK: fetchItemsByHTTPMethod [String: Any]
+    /**
+     Fetch items with HTTP method.
+     
+     Fetch items with HTTP method with body parameter. And decodes the data with provided `Decodable` model. It's extreamly handy if anyone just  want to provide a data model and url and get back the decoded data. Uses swift concurrency.
+     
+     - Parameters:
+        - urlLink: Receives an `Optional<URL>` aka `URL?`
+        - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
+        - body: `Optional<[String: Any]>` aka `[String: Any]?` for sending to remote server.
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
+     
+     - Returns: Returns a  `Result<Success, Failure>` type where `Success` is  ``NetworkResult`` and failure is `Error`
+     */
+    public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: [String: Any]?, additionalHeader: [Header]? = nil) async throws -> NetworkResult<Data> {
+        let uploadData = RKAPIHelper.generateRequestBody(body)
+        
+        let reply = try await fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: uploadData, additionalHeader: additionalHeader)
+        
+        return reply
+    }
+    
+    /**
+     Fetch items with HTTP method.
+     
+     Fetch items with HTTP method with body parameter. And decodes the data with provided `Decodable` model. It's extreamly handy if anyone just  want to provide a data model and url and get back the decoded data. Uses swift concurrency.
+     
+     - Parameters:
+        - urlLink: Receives an `Optional<URL>` aka `URL?`
+        - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
+        - body: `Optional<[String: Any]>` aka `[String: Any]?` for sending to remote server.
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
+        - model: Generic Type `D` where `D` confirms to `Decodable`
+     
+     - Returns: Returns a  `Result<Success, Failure>` type where `Success` is  ``NetworkResult`` and failure is `Error`
+     */
+    public func fetchItemsByHTTPMethod<D: Decodable>(urlLink: URL?, httpMethod: HTTPMethod, body: [String: Any]?, additionalHeader: [Header]? = nil, _ model: D.Type) async -> Result<NetworkResult<D>, Error> {
+        let uploadData = RKAPIHelper.generateRequestBody(body)
+        
+        do {
+            let reply = try await fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: uploadData, additionalHeader: additionalHeader)
+            
+            guard let rawData = reply.data else {throw reply.response}
+            
+            let decodedData = try JSONDecoder().decode(model.self, from: rawData)
+            
+            return .success(NetworkResult(data: decodedData, response: reply.response))
+        } catch {
+            return .failure(error)
+        }
+    }
 }
 
-//MARK: With Combine Publisher
+//MARK: - With Combine Publisher
 @available(iOS 13.0, macOS 10.15.0, *)
 extension RKAPIService {
     /**
@@ -255,11 +347,11 @@ extension RKAPIService {
      
      - Parameters:
         - urlLink: Receives an `Optional<URL>` aka `URL?`
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
      
      - Returns: Returns a  `AnyPublisher<Success, Failure>` where `Success` is ``NetworkResult`` `Failure` is `Error`
      */
-    public func fetchItems(urlLink: URL?, additionalHeader: [HTTPHeader]? = nil) -> AnyPublisher<NetworkResult<Data>, Error> {
+    public func fetchItems(urlLink: URL?, additionalHeader: [Header]? = nil) -> AnyPublisher<NetworkResult<Data>, Error> {
         guard let url = urlLink else {
             return Fail(error: URLError(.badURL)).eraseToAnyPublisher()
         }
@@ -296,12 +388,12 @@ extension RKAPIService {
      
      - Parameters:
         - urlLink: Receives an `Optional<URL>` aka `URL?`
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - model: Generic Type `D` where `D` confirms to `Decodable`
      
      - Returns: Returns a  `AnyPublisher<Success, Failure>` where `Success` is ``NetworkResult`` `Failure` is `Error`
      */
-    public func fetchItems<D: Decodable>(urlLink: URL?, additionalHeader: [HTTPHeader]? = nil, _ model: D.Type) -> AnyPublisher<NetworkResult<D>, Error> {
+    public func fetchItems<D: Decodable>(urlLink: URL?, additionalHeader: [Header]? = nil, _ model: D.Type) -> AnyPublisher<NetworkResult<D>, Error> {
         return fetchItems(urlLink: urlLink, additionalHeader: additionalHeader)
             .tryMap{ reply in
                 guard let rawData = reply.data else {throw reply.response}
@@ -326,11 +418,11 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: `Optional<Data>` aka `Data?` for sending to remote server.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
      
      - Returns: Returns a  `AnyPublisher<Success, Failure>` where Success is ``NetworkResult`` Failure is `Error`
      */
-    public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, additionalHeader: [HTTPHeader]? = nil) -> AnyPublisher<NetworkResult<Data>, Error> {
+    public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, additionalHeader: [Header]? = nil) -> AnyPublisher<NetworkResult<Data>, Error> {
         guard let url = urlLink else {
             return Fail(error: URLError(.badURL)).eraseToAnyPublisher()
         }
@@ -358,9 +450,9 @@ extension RKAPIService {
                 guard let response = output.response as? HTTPURLResponse else {
                     throw URLError(.cannotParseResponse)
                 }
-
+                
                 let status = HTTPStatusCode(rawValue: response.statusCode)
-
+                
                 return NetworkResult(data: output.data, response: status)
             }
             .eraseToAnyPublisher()
@@ -375,7 +467,7 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: `Optional<Data>` aka `Data?` for sending to remote server.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - model: Generic Type `D` where `D` confirms to `Decodable`
      
      - Returns: Returns a  `AnyPublisher<Success, Failure>` where Success is ``NetworkResult`` Failure is `Error`
@@ -383,7 +475,7 @@ extension RKAPIService {
     public func fetchItemsByHTTPMethod<D: Decodable>(urlLink: URL?,
                                                      httpMethod: HTTPMethod,
                                                      body: Data?,
-                                                     additionalHeader: [HTTPHeader]? = nil,
+                                                     additionalHeader: [Header]? = nil,
                                                      _ model: D.Type) -> AnyPublisher<NetworkResult<D>, Error> {
         return fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: body, additionalHeader: additionalHeader)
             .tryMap{ reply in
@@ -408,15 +500,51 @@ extension RKAPIService {
      - Parameters:
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
+        - body: `Optional<Data>` aka `Data?` for sending to remote server.
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
+        - model: Generic Type `D` where `D` confirms to `Decodable`
+     
+     - Returns: Returns a  `AnyPublisher<Success, Failure>` where Success is ``NetworkResult`` Failure is `Error`
+     */
+    public func fetchItemsByHTTPMethod<D: Decodable>(urlLink: URL?,
+                                                     httpMethod: HTTPMethod,
+                                                     body: [String: Any]?,
+                                                     additionalHeader: [Header]? = nil,
+                                                     _ model: D.Type) -> AnyPublisher<NetworkResult<D>, Error> {
+        let uploadData = RKAPIHelper.generateRequestBody(body)
+        
+        return fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: uploadData, additionalHeader: additionalHeader)
+            .tryMap{ reply in
+                guard let rawData = reply.data else {throw reply.response}
+                
+                let decodedData = try JSONDecoder().decode(model.self, from: rawData)
+                
+                return NetworkResult(data: decodedData, response: reply.response)
+            }
+            .mapError{ error in
+                
+                return error
+            }
+            .eraseToAnyPublisher()
+    }
+    
+    /**
+     Fetch items with HTTP method.
+     
+     Fetch items with HTTP method with body parameter. And decodes the data with provided `Decodable` model. It's extreamly handy if anyone just  want to provide a data model and url and get back the decoded data. Uses Combine Publisher.
+     
+     - Parameters:
+        - urlLink: Receives an `Optional<URL>` aka `URL?`
+        - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: Generic Type `E` where `E` confirms to `Encodable`.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
      
      - Returns: Returns a  `AnyPublisher<Success, Failure>` where Success is ``NetworkResult`` Failure is `Error`
      */
     public func fetchItemsByHTTPMethod<E: Encodable>(urlLink: URL?,
                                                      httpMethod: HTTPMethod,
                                                      body: E,
-                                                     additionalHeader: [HTTPHeader]? = nil) -> AnyPublisher<NetworkResult<Data>, Error> {
+                                                     additionalHeader: [Header]? = nil) -> AnyPublisher<NetworkResult<Data>, Error> {
         let uploadData = RKAPIHelper.generateRequestBody(body)
         
         return fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: uploadData, additionalHeader: additionalHeader)
@@ -431,7 +559,7 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: Generic Type `E` where `E` confirms to `Encodable`.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - model: Generic Type `D` where `D` confirms to `Decodable`
      
      - Returns: Returns a  `AnyPublisher<Success, Failure>` where Success is ``NetworkResult`` Failure is `Error`
@@ -439,14 +567,41 @@ extension RKAPIService {
     public func fetchItemsByHTTPMethod<D: Decodable, E: Encodable>(urlLink: URL?,
                                                                    httpMethod: HTTPMethod,
                                                                    body: E,
-                                                                   additionalHeader: [HTTPHeader]? = nil,
+                                                                   additionalHeader: [Header]? = nil,
                                                                    _ model: D.Type) -> AnyPublisher<NetworkResult<D>, Error> {
+        let uploadData = RKAPIHelper.generateRequestBody(body)
+        
+        return fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: uploadData, additionalHeader: additionalHeader, D.self)
+    }
+    
+    //MARK: fetchItemsByHTTPMethod [String: Any]
+    
+    /**
+     Fetch items with HTTP method.
+     
+     Fetch items with HTTP method with body parameter. And decodes the data with provided `Decodable` model. It's extreamly handy if anyone just  want to provide a data model and url and get back the decoded data. Uses Combine Publisher.
+     
+     - Parameters:
+        - urlLink: Receives an `Optional<URL>` aka `URL?`
+        - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
+        - body: Generic Type `E` where `E` confirms to `Encodable`.
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
+        - model: Generic Type `D` where `D` confirms to `Decodable`
+     
+     - Returns: Returns a  `AnyPublisher<Success, Failure>` where Success is ``NetworkResult`` Failure is `Error`
+     */
+    public func fetchItemsByHTTPMethod<D: Decodable>(urlLink: URL?,
+                                                     httpMethod: HTTPMethod,
+                                                     body: [String: Any],
+                                                     additionalHeader: [Header]? = nil,
+                                                     _ model: D.Type) -> AnyPublisher<NetworkResult<D>, Error> {
         let uploadData = RKAPIHelper.generateRequestBody(body)
         
         return fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: uploadData, additionalHeader: additionalHeader, D.self)
     }
 }
 
+//MARK: - Completion Handller
 @available(iOS, deprecated: 13.0, obsoleted: 14.8, message: "Completion handler may occur memory leak, user async/await or Combine Publisher instead")
 @available(macOS, deprecated: 10.15.0, obsoleted: 11.6.7, message: "Completion handler may occur memory leak, user async/await or Combine Publisher instead")
 extension RKAPIService {
@@ -457,10 +612,10 @@ extension RKAPIService {
      
      - Parameters:
         - urlLink: Receives an `Optional<URL>` aka `URL?`
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - completion: An `@escaping` closure parameter which provides a `Result<Success, Failure>` where `Success` is ``NetworkResult`` and `Failure` is `Error` as return of closure
      */
-    public func fetchItems(urlLink: URL?, additionalHeader: [HTTPHeader]? = nil, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void ){
+    public func fetchItems(urlLink: URL?, additionalHeader: [Header]? = nil, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void ){
         guard let url = urlLink else {
             completion(.failure(URLError(.badURL)))
             
@@ -501,12 +656,12 @@ extension RKAPIService {
      
      - Parameters:
         - urlLink: Receives an `Optional<URL>` aka `URL?`
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - model: Generic Type `D` where `D` confirms to `Decodable`
         - completion: An `@escaping` closure parameter which provides a `Result<Success, Failure>` where `Success` is ``NetworkResult`` and `Failure` is `Error` as return of closure
      */
     public func fetchItems<D: Decodable>(urlLink: URL?,
-                                         additionalHeader: [HTTPHeader]? = nil,
+                                         additionalHeader: [Header]? = nil,
                                          _ model: D.Type,
                                          _ completion: @escaping (Result<NetworkResult<D>, Error>)-> Void ){
         fetchItems(urlLink: urlLink, additionalHeader: additionalHeader) { result in
@@ -537,10 +692,10 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: `Optional<Data>` aka `Data?` for sending to remote server.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - completion: An `@escaping` closure parameter which provides a `Result<Success, Failure>` where `Success` is ``NetworkResult`` and `Failure` is `Error` as return of closure
      */
-    public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, additionalHeader: [HTTPHeader]? = nil, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void ){
+    public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, additionalHeader: [Header]? = nil, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void ){
         guard let url = urlLink else {
             completion(.failure(URLError(.badURL)))
             
@@ -589,14 +744,14 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: `Optional<Data>` aka `Data?` for sending to remote server.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - model: Generic Type `D` where `D` confirms to `Decodable`
         - completion: An `@escaping` closure parameter which provides a `Result<Success, Failure>` where `Success` is ``NetworkResult`` and `Failure` is `Error` as return of closure
      */
     public func fetchItemsByHTTPMethod<D: Decodable>(urlLink: URL?,
                                                      httpMethod: HTTPMethod,
                                                      body: Data?,
-                                                     additionalHeader: [HTTPHeader]? = nil,
+                                                     additionalHeader: [Header]? = nil,
                                                      _ model: D.Type,
                                                      _ completion: @escaping (Result<NetworkResult<D>, Error>)-> Void) {
         fetchItemsByHTTPMethod(urlLink: urlLink, httpMethod: httpMethod, body: body, additionalHeader: additionalHeader) { result in
@@ -627,13 +782,13 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: Generic Type `E` where `E` confirms to `Encodable`.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - completion: An `@escaping` closure parameter which provides a `Result<Success, Failure>` where `Success` is ``NetworkResult`` and `Failure` is `Error` as return of closure
      */
     public func fetchItemsByHTTPMethod<E: Encodable>(urlLink: URL?,
                                                      httpMethod: HTTPMethod,
                                                      body: E,
-                                                     additionalHeader: [HTTPHeader]? = nil,
+                                                     additionalHeader: [Header]? = nil,
                                                      _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void) {
         let uploadData = RKAPIHelper.generateRequestBody(body)
         
@@ -649,14 +804,14 @@ extension RKAPIService {
         - urlLink: Receives an `Optional<URL>` aka `URL?`
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: Generic Type `E` where `E` confirms to `Encodable`.
-        - additionalHeader: Receives an `Optional<Array<HTTPHeader>>` aka [``HTTPHeader``]?
+        - additionalHeader: Receives an `Optional<Array<Header>>` aka [``Header``]?
         - model: Generic Type `D` where `D` confirms to `Decodable`
         - completion: An `@escaping` closure parameter which provides a `Result<Success, Failure>` where `Success` is ``NetworkResult`` and `Failure` is `Error` as return of closure
      */
     public func fetchItemsByHTTPMethod<D: Decodable, E: Encodable>(urlLink: URL?,
                                                                    httpMethod: HTTPMethod,
                                                                    body: E,
-                                                                   additionalHeader: [HTTPHeader]? = nil,
+                                                                   additionalHeader: [Header]? = nil,
                                                                    _ model: D.Type,
                                                                    _ completion: @escaping (Result<NetworkResult<D>, Error>)-> Void) {
         let uploadData = RKAPIHelper.generateRequestBody(body)