+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0](https://github.com/TheRakiburKhan/RKAPIService/releases/tag/2.5.0) - 16 October 2022
+### Added
+- `Authorization` enum added for HTTP Authorization. It can be used in `additionalHeader` parameter of a request.
+- Added support for Dictionary type `[String: Any]` as request body
+
+### Removed
+- `RKAPIServiceProtocol` is removed
+
+### Fixed
+- `HTTPHeader` inaccessible bug fix
 
 ## [2.4.1](https://github.com/TheRakiburKhan/RKAPIService/releases/tag/2.4.1) - 10 August 2022
 ### Changed